+++ core/js/src/main/scala/org/http4s/util/PlatformUtil.scala
@@ -1,0 +1,4 @@
+package org.http4s
+package util
+
+trait PlatformUtil
