+++ core/jvm/src/main/scala/org/http4s/PlatformEntityDecoder.scala
@@ -2,7 +2,6 @@ package org.http4s
 
 import java.io.{File, FileOutputStream, PrintStream}
 
-import cats._
 import cats.effect.Sync
 import fs2.io._
 
@@ -13,17 +12,16 @@ trait PlatformEntityDecoderInstances {
   /////////////////// Instances //////////////////////////////////////////////
 
   // File operations // TODO: rewrite these using NIO non blocking FileChannels, and do these make sense as a 'decoder'?
-  def binFile[F[_]: MonadError[?[_], Throwable]](file: File)(
-      implicit F: Sync[F]): EntityDecoder[F, File] =
+  def binFile[F[_]](file: File)(implicit F: Sync[F]): EntityDecoder[F, File] =
     EntityDecoder.decodeBy(MediaRange.`*/*`) { msg =>
       val sink = writeOutputStream[F](F.delay(new FileOutputStream(file)))
       DecodeResult.success(msg.body.to(sink).run).map(_ => file)
     }
 
-  def textFile[F[_]: MonadError[?[_], Throwable]](file: File)(
-      implicit F: Sync[F]): EntityDecoder[F, File] =
+  def textFile[F[_]](file: File)(implicit F: Sync[F]): EntityDecoder[F, File] =
     EntityDecoder.decodeBy(MediaRange.`text/*`) { msg =>
       val sink = writeOutputStream[F](F.delay(new PrintStream(new FileOutputStream(file))))
       DecodeResult.success(msg.body.to(sink).run).map(_ => file)
     }
+
 }
