+++ core/shared/src/main/scala/org/http4s/EntityDecoder.scala
@@ -214,19 +214,6 @@ trait EntityDecoderInstances extends PlatformEntityDecoderInstances {
   implicit def charArrayDecoder[F[_]: Sync]: EntityDecoder[F, Array[Char]] =
     text.map(_.toArray)
 
-  // File operations // TODO: rewrite these using NIO non blocking FileChannels, and do these make sense as a 'decoder'?
-  def binFile[F[_]](file: File)(implicit F: Sync[F]): EntityDecoder[F, File] =
-    EntityDecoder.decodeBy(MediaRange.`*/*`) { msg =>
-      val sink = writeOutputStream[F](F.delay(new FileOutputStream(file)))
-      DecodeResult.success(msg.body.to(sink).run).map(_ => file)
-    }
-
-  def textFile[F[_]](file: File)(implicit F: Sync[F]): EntityDecoder[F, File] =
-    EntityDecoder.decodeBy(MediaRange.`text/*`) { msg =>
-      val sink = writeOutputStream[F](F.delay(new PrintStream(new FileOutputStream(file))))
-      DecodeResult.success(msg.body.to(sink).run).map(_ => file)
-    }
-
   implicit def multipart[F[_]: Sync]: EntityDecoder[F, Multipart[F]] =
     MultipartDecoder.decoder
