+++ core/shared/src/main/scala/org/http4s/internal/Macros.scala
