+++ core/shared/src/main/scala/org/http4s/syntax/LiteralsSyntax.scala
