+++ core/shared/src/main/scala/org/http4s/syntax/NonEmptyListSyntax.scala
