+++ testing/shared/src/main/scala/org/http4s/testing/EntityEncoderTests.scala
