+++ tests/shared/src/test/scala/org/http4s/EntityEncoderSpec.scala
@@ -9,7 +9,6 @@ import cats.laws.discipline.ContravariantTests
 import cats.laws.discipline.eq._
 import fs2._
 import java.io._
-import java.nio.charset.StandardCharsets
 import java.util.concurrent.TimeoutException
 import org.http4s.headers._
 import org.scalacheck.Arbitrary
