+++ tests/shared/src/test/scala/org/http4s/SchemeSpec.scala
