+++ tests/shared/src/test/scala/org/http4s/headers/ContentEncodingSpec.scala
